pyproject.toml
[project]
name = "omnidocbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apted>=1.0.3",
    "beautifulsoup4>=4.13.3",
    "evaluate>=0.4.3",
    "func-timeout>=4.3.5",
    "loguru>=0.7.3",
    "lxml>=5.3.1",
    "matplotlib>=3.10.1",
    "mmeval>=0.2.1",
    "nltk>=3.9.1",
    "pylatexenc>=2.10",
    "python-levenshtein>=0.27.1",
    "pyyaml>=6.0.2",
    "scipy>=1.15.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[tool.ruff]
ignore = [
    "F401",  # Ignore 'imported but unused' errors
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
